--- src/main/java/com/newsfeed/testagram/domain/login/service/CustomUserDetailsService.java
@@ -1,32 +1,0 @@
-package com.newsfeed.testagram.domain.login.service;
-
-import com.newsfeed.testagram.domain.member.entity.Member;
-import com.newsfeed.testagram.domain.member.repository.MemberRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.Collections;
-
-@Service
-@RequiredArgsConstructor
-public class CustomUserDetailsService implements UserDetailsService {
-
-    private final MemberRepository memberRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Member member = memberRepository.findByEmail(username)
-                .or(() -> memberRepository.findByNickname(username))
-                .orElseThrow(() -> new UsernameNotFoundException("해당 사용자가 존재하지 않습니다."));
-
-        return new User(
-                member.getEmail(),
-                member.getPassword(),
-                Collections.emptyList()
-        );
-    }
-}
